+++ resources/aws_rds_db_proxy_target_group.rb
@@ -67,5 +67,5 @@ rest_property_map({
 })
 
 rest_post_only_properties %i{
-  db_proxy_name target_group_name
+  target_group_name
 }